@@ -20,6 +20,8 @@ namespace APSIM.Shared.Documentation
 
         private static Dictionary<Assembly, XmlDocument> documentCache = new Dictionary<Assembly, XmlDocument>();
 
+        private static Dictionary<string, string> modelDescriptionCache = new Dictionary<string, string>();
+
         /// <summary>
         /// Get the summary of a type removing CRLF.
         /// </summary>
@@ -85,42 +87,87 @@ namespace APSIM.Shared.Documentation
         {
             XmlDocument document = LoadDocument(member.DeclaringType.Assembly);
             
-            string result = "";
+            //make a list of types to check
+            Type type = member.ReflectedType.BaseType;
             List<Type> types = new List<Type>();
-            types.Add(member.ReflectedType);
-
-            while (types.Count > 0 && string.IsNullOrEmpty(result))
+            while(type != null)
             {
-                Type type = types.Last();
-                types.Remove(type);
+                types.Add(type);
+                type = type.BaseType;
+            }
+            types.Add(member.ReflectedType);
+            //Then reverse the list to do this type first, then work upwards from model for preformance
+            types.Reverse();
 
-                if (type.BaseType != null)
-                    types.Add(type.BaseType);
-
-                foreach (Type t in type.GetInterfaces())
-                    types.Add(t);                
-
-                string fullName = type.FullName + "." + member.Name;
-
-                if (member is PropertyInfo)
-                    result = GetDocumentationElement(document, fullName, tagName, 'P');
-                else if (member is FieldInfo)
-                    result = GetDocumentationElement(document, fullName, tagName, 'F');
-                else if (member is EventInfo)
-                    result = GetDocumentationElement(document, fullName, tagName, 'E');
-                else if (member is MethodInfo method)
-                {
-                    string args = string.Join(",", method.GetParameters().Select(p => p.ParameterType.FullName));
-                    args = args.Replace("+", ".");
-                    result = GetDocumentationElement(document, $"{fullName}({args})", tagName, 'M');
-                }
-                else
-                {
-                    throw new ArgumentException($"Unknown argument type {member.GetType().Name}");
-                }
+            foreach(Type t in types)
+            {
+                string result = GetCustomTagDetails(member, tagName, document, t);
+                if (!string.IsNullOrEmpty(result))
+                    return result;
             }
 
-            return result;
+            return "";
+        }
+
+        /// <summary>
+        /// 
+        /// </summary>
+        /// <param name="member">The member.</param>
+        /// <param name="tagName">The name of the xml element in the documentation to be reaed. E.g. "summary".</param>
+        /// <param name="document"></param>
+        /// <param name="type"></param>
+        public static string GetCustomTagDetails(MemberInfo member, string tagName, XmlDocument document, Type type)
+        {
+            string fullName = type.FullName + "." + member.Name;
+
+            char typeLetter = ' ';
+            if (member is PropertyInfo)
+                typeLetter = 'P';
+            else if (member is FieldInfo)
+                typeLetter = 'F';
+            else if (member is EventInfo)
+                typeLetter = 'E';
+            else if (member is MethodInfo method)
+                typeLetter = 'M';
+
+            if (typeLetter == 'P' || typeLetter == 'F' || typeLetter == 'E')
+            {
+                string cacheName = fullName+tagName;
+                string description = "";
+                if (modelDescriptionCache.TryGetValue(cacheName, out description))
+                    return description;
+
+                description = GetDocumentationElement(document, fullName, tagName, typeLetter);
+                if (!string.IsNullOrEmpty(description))
+                {
+                    modelDescriptionCache.Add(cacheName, description);
+                    return description;
+                }
+            }
+            else if (typeLetter == 'M')
+            {
+                MethodInfo method = member as MethodInfo;
+                string args = string.Join(",", method.GetParameters().Select(p => p.ParameterType.FullName));
+                args = args.Replace("+", ".");
+                string name = $"{fullName}({args})";
+
+                string cacheName = name+tagName;
+                string description = "";
+                if (modelDescriptionCache.TryGetValue(cacheName, out description))
+                    return description;
+
+                description = GetDocumentationElement(document, name, tagName, typeLetter);
+                if (!string.IsNullOrEmpty(description))
+                {
+                    modelDescriptionCache.Add(cacheName, description);
+                    return description;
+                }
+            }
+            else
+            {
+                throw new ArgumentException($"Unknown argument type {member.GetType().Name}");
+            }
+            return "";
         }
 
         /// <summary>
