@@ -192,11 +192,7 @@ namespace UserInterface.Presenters
         {
             try
             {
-                manager.Reformat();
-                if (managerView.Editor != null)
-                {
-                    managerView.Editor.Text = manager.Code;
-                }
+                managerView.Editor.Text = CodeFormatting.Reformat(managerView.Editor.Text);
             }
             catch (Exception err)
             {
