@@ -1,9 +1,9 @@
 ﻿using System;
 using System.Collections.Generic;
-using System.Data;
 using UserInterface.Interfaces;
 using Gtk;
-using UserInterface.Presenters;
+using System.IO;
+using APSIM.Shared.Utilities;
 
 // This is the view used by the WeatherFile component
 namespace UserInterface.Views
@@ -46,6 +46,7 @@ namespace UserInterface.Views
         public event BrowseDelegate ConstantsFileSelected;
 
         private Label labelFileName = null;
+        private Label labelFileNameRelative = null;
         private Box vbox1 = null;
         private Notebook notebook1 = null;
         private TextView textview1 = null;
@@ -72,6 +73,7 @@ namespace UserInterface.Views
         {
             Builder builder = BuilderFromResource("ApsimNG.Resources.Glade.TabbedMetDataView.glade");
             labelFileName = (Label)builder.GetObject("labelFileName");
+            labelFileNameRelative = (Label)builder.GetObject("labelFileNameRelative");
             vbox1 = (Box)builder.GetObject("vbox1");
             notebook1 = (Notebook)builder.GetObject("notebook1");
             textview1 = (TextView)builder.GetObject("textview1");
@@ -145,6 +147,22 @@ namespace UserInterface.Views
         {
             get { return labelFileName.Text; }
             set { labelFileName.Text = value; }
+        }
+
+        /// <summary>Gets or sets the filename.</summary>
+        /// <value>The filename.</value>
+        public string FilenameRelative
+        {
+            set { 
+                if (value.StartsWith("%root%"))
+                    labelFileNameRelative.Text = value;
+                else if (value.CompareTo(PathUtilities.GetAbsolutePath(value, null)) == 0)
+                    labelFileNameRelative.Text = value;
+                else if (value.StartsWith('/') || value.StartsWith('\\'))
+                    labelFileNameRelative.Text = "." + value;
+                else
+                    labelFileNameRelative.Text = "./" + value;
+            }
         }
 
         /// <summary>Gets or sets the filename.</summary>
@@ -481,6 +499,9 @@ namespace UserInterface.Views
 
         /// <summary>Gets or sets the filename.</summary>
         string Filename { get; set; }
+
+        /// <summary>Gets or sets the filename.</summary>
+        string FilenameRelative { set; }
 
         /// <summary>Gets or sets the filename.</summary>
         string ConstantsFileName { get; set; }
