@@ -1,36 +1,71 @@
 ﻿using System;
+using System.Linq;
+using Models.Core;
+using Models.Functions;
 
 namespace Models
 {
     /// <summary>A class for holding a fertiliser type.</summary>
     [Serializable]
-    public class FertiliserType
+    [PresenterName("UserInterface.Presenters.PropertyPresenter")]
+    [ViewName("UserInterface.Views.PropertyView")]
+    [ValidParent(ParentType = typeof(Fertiliser))]
+    public class FertiliserType : Model
     {
-        /// <summary>The name of the fertiliser type.</summary>
-        public string Name { get; set; }
-
         /// <summary>A description of the fertiliser type.</summary>
+        [Description("Description")]
         public string Description { get; set; }
 
-        /// <summary>The fraction of no3.</summary>
-        public double FractionNO3 { get; set; }
+        /// <summary>The fraction of fertiliser at which the remainder is released (e.g. when 0.995 released)</summary>
+        [Description("The fraction of fertiliser at which the remainder is released (e.g. when 0.995 released)")]
+        public double FractionWhenRemainderReleased { get; set; } = 0.995;
 
-        /// <summary>The fraction of nh4.</summary>
-        public double FractionNH4 { get; set; }
+        /// <summary>The name of solute 1.</summary>
+        [Description("Solute 1 name")]
+        public string Solute1Name { get; set; }
 
-        /// <summary>The fraction of urea.</summary>
-        public double FractionUrea { get; set; }
+        /// <summary>Solute 1 fraction.</summary>
+        [Description("Fraction of solute 1 in fertiliser")]
+        public double Solute1Fraction { get; set; }
 
-        /// <summary>The fraction of rock p.</summary>
-        public double FractionRockP { get; set; }
+        /// <summary>The name of solute 2.</summary>
+        [Description("Solute 2 name")]
+        public string Solute2Name { get; set; }
 
-        /// <summary>The fraction of banded p.</summary>
-        public double FractionBandedP { get; set; }
+        /// <summary>Solute 2 fraction.</summary>
+        [Description("Fraction of solute 2 in fertiliser")]
+        public double Solute2Fraction { get; set; }
 
-        /// <summary>The fraction of labile p.</summary>
-        public double FractionLabileP { get; set; }
+        /// <summary>The name of solute 3.</summary>
+        [Description("Solute 3 name")]
+        public string Solute3Name { get; set; }
 
-        /// <summary>The fraction of ca.</summary>
-        public double FractionCa { get; set; }
+        /// <summary>Solute 3 fraction.</summary>
+        [Description("Fraction of solute 3 in fertiliser")]
+        public double Solute3Fraction { get; set; }
+
+        /// <summary>The name of solute 4.</summary>
+        [Description("Solute 4 name")]
+        public string Solute4Name { get; set; }
+
+        /// <summary>Solute 4 fraction.</summary>
+        [Description("Fraction of solute 4 in fertiliser")]
+        public double Solute4Fraction { get; set; }
+
+        /// <summary>The name of solute 5.</summary>
+        [Description("Solute 5 name")]
+        public string Solute5Name { get; set; }
+
+        /// <summary>Solute 5 fraction.</summary>
+        [Description("Fraction of solute 5 in fertiliser")]
+        public double Solute5Fraction { get; set; }
+
+        /// <summary>The name of solute 6.</summary>
+        [Description("Solute 6 name")]
+        public string Solute6Name { get; set; }
+
+        /// <summary>Solute 6 fraction.</summary>
+        [Description("Fraction of solute 6 in fertiliser")]
+        public double Solute6Fraction { get; set; }
     }
 }
