@@ -19,12 +19,26 @@ namespace Models.PMF.Phen
         [Link]
         Phenology Phenology = null;
 
-        /// <summary>Haun stage is used for zadok stages 10 to 30</summary>
-        [Link(Type = LinkType.Path, Path = "[Phenology].HaunStage")]
-        IFunction haunStage = null;
-
         [Link]
         private IPlant plant = null;
+
+        /// <summary>The thermal time</summary>
+        [Link(Type = LinkType.Child, ByName = true)]
+        public IFunction TillerNumber = null;
+
+        /// <summary>The thermal time</summary>
+        [Link(Type = LinkType.Child, ByName = true)]
+        public IFunction haunStage = null;
+
+        /// <summary>
+        /// Zadok stage numbers for wheat
+        /// </summary>
+        public static readonly double[] ZADOK_STAGE_NUMBERS = [30.0, 34, 39.0, 55.0, 65.0, 71.0, 87.0, 90.0];
+
+        /// <summary>
+        /// Growth stage numbers for wheat
+        /// </summary>
+        public static readonly double[] GROWTH_STAGE_NUMBERS = [5.0, 5.99, 6.0, 7.0, 8.0, 9.0, 10.0, 11.0];
 
         /// <summary>Gets the stage.</summary>
         /// <value>The stage.</value>
@@ -47,11 +61,9 @@ namespace Models.PMF.Phen
                 }
                 else if (!Phenology.InPhase("ReadyForHarvesting"))
                 {
-                    double[] zadok_code_y = { 30.0, 34, 39.0, 55.0, 65.0, 71.0, 87.0, 90.0 };
-                    double[] zadok_code_x = { 5.0, 5.99, 6.0, 7.0, 8.0, 9.0, 10.0, 11.0 };
                     bool DidInterpolate;
                     zadok_stage = MathUtilities.LinearInterpReal(Phenology.Stage,
-                                                               zadok_code_x, zadok_code_y,
+                                                               GROWTH_STAGE_NUMBERS, ZADOK_STAGE_NUMBERS,
                                                                out DidInterpolate);
                 }
                 else if (Phenology.InPhase("ReadyForHarvesting"))
